@@ -1,6 +1,6 @@
 module github.com/artie-labs/transfer
 
-go 1.23.0
+go 1.24.0
 
 require (
 	cloud.google.com/go/bigquery v1.65.0
