@@ -1,7 +1,6 @@
 package kafkalib
 
 import (
-	"context"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -27,7 +26,7 @@ func TestConnection_Mechanism(t *testing.T) {
 }
 
 func TestConnection_Dialer(t *testing.T) {
-	ctx := context.Background()
+	ctx := t.Context()
 	{
 		// Plain
 		c := NewConnection(false, false, "", "")
